+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Version 0.2.0-SNAPSHOT
 -------------
 * `keepUnused = true` keeps unused version declarations too now
+* Plugins are no longer moved from the libraries section so that a library declaration for a plugin can still be updated.
 
 Version 0.1.0
 -------------